fixed compiler warning
git-svn-id: http://svn.berlios.de/svnroot/repos/sipsak/trunk@338 75b5f7c7-cfd4-0310-b54c-e118b2c5249a

+++ sipsak/header_f.c
@@ -554,5 +554,5 @@ void print_message_line(char *message)
 	}
 	else if (*(crlf - 1) == '\r')
 		crlf--;
-	printf("%.*s\n", crlf - message, message);
+	printf("%.*s\n", (int)(crlf - message), message);
 }
